+++ lib/flutter_crop_camera_windows.dart
@@ -1,0 +1,10 @@
+import 'flutter_crop_camera_platform_interface.dart';
+
+class FlutterCropCameraWindows extends FlutterCropCameraPlatform {
+  static void registerWith() {
+    FlutterCropCameraPlatform.instance = FlutterCropCameraWindows();
+  }
+
+  @override
+  Future<String?> getPlatformVersion() async => 'Windows';
+}